Tested BFS and DFS.  Need to add a gui
