+++ src/org/BFS.java
@@ -61,32 +61,6 @@ public class BFS extends PathfindingStrategy{
     throw new NoPathException();
   }
 
-  /**
-   * Helper function that iterates through Points using parents. Starts
-   * at the goal and moves to the starting Point. Reverses this order when
-   * complete.
-   * @param start: starting Point
-   * @param goal: ending Point
-   * @return the path from the starting Point to the ending Point
-   */
-  public ArrayList<Point> findPath(Point start, Point goal){
 
-    // Initializes the path to be returned
-    ArrayList<Point> path = new ArrayList<Point>();
-    path.add(goal);
-    // Sets the current Point to iterate through.
-    Point currPoint = goal;
-
-    // Runs though the path by starting at the end and using parents to
-    // find the next position up.
-    while(currPoint != start){
-      path.add(currPoint.parent);
-      currPoint = currPoint.getParent();
-    }
-
-    // Reverses the path so it moves from start to end instead of end to start
-    Collections.reverse(path);
-    return path;
-  }
 
 }
