+++ src/org/DFS.java
@@ -1,7 +1,9 @@
 package org;
 
+import com.sun.org.apache.bcel.internal.generic.NOP;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.Stack;
 
 /**
  * Created by Tom on 4/15/2017.
@@ -10,9 +12,10 @@ public class DFS extends PathfindingStrategy {
 
   HashSet<Point> visited = new HashSet<>(); // Used to prevent stack overflow
   Point startingLoc; // Used to validate where we are in recursion
+  Stack<Point> stack;
 
   /**
-   * Recursively find the first path that connects start to goal using a DFS algorithm
+   * uses a stack to find the first path that connects start to goal using a DFS algorithm
    * @implNote this completely ignores any heuristics
    * @param start: starting Point
    * @param goal: ending Point
@@ -20,34 +23,37 @@ public class DFS extends PathfindingStrategy {
    * @throws NoPathException
    */
   @Override
-  public ArrayList<Point> execute(Point start, Point goal) throws NoPathException {
+  public ArrayList<Point> execute(Point start, Point goal) throws NoPathException{
+    stack = new Stack();
+    visited = new HashSet<>();
+    Point currPoint;
+
+    // Used to track visited nodes.
     visited.add(start);
+    // Used to queue through nodes to check and add neighbors.
+    stack.push(start);
 
-    if(startingLoc == null){
-      startingLoc = start;
-    }
+    while (stack.size() != 0) {
+      //Sets the current point in the queue by retrieving and removing the list's head.
+      currPoint = stack.pop();
 
-    if(start.equals(goal)){
-      return ListPath(goal, startingLoc);
-    }
-
-    for(Point neighbor: start.getNeighbors()){
-      if(visited.contains(neighbor)) {
-        continue;
+      // If a neighbor has not yet been visited add
+      // them to the queue and to visited. Additionally,
+      // set their parent to the current node.
+      for (Point neighbor : currPoint.getNeighbors()) {
+        if (!visited.contains(neighbor)) {
+          neighbor.addParent(currPoint);
+          stack.push(neighbor);
+          visited.add(neighbor);
+        }
       }
 
-      neighbor.parent = start;
-      ArrayList<Point> result =  execute(neighbor, goal);
-      if(result != null){
-        return result;
+      // If the current point is the goal, use the helper function to return the path
+      // and clear the queue and visited lists.
+      if (currPoint == goal) {
+        return findPath(start, goal);
       }
     }
-
-    if(start.equals(startingLoc)) {
-      startingLoc = null;
-      throw new NoPathException();
-    }else {
-      return null;
-    }
+    throw new NoPathException();
   }
 }
